pyproject.toml
[project]
name = "paper_helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.1.3",
    "beautifulsoup4>=4.13.3",
    "cachetools>=5.5.2",
    "faiss-cpu>=1.10.0",
    "fastapi>=0.115.12",
    "flask>=3.1.0",
    "icecream>=2.1.4",
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.20",
    "langchain-openai>=0.3.11",
    "lxml>=5.3.1",
    "openai>=1.69.0",
    "pdfplumber>=0.11.6",
    "pydantic>=2.11.0",
    "pymupdf>=1.25.4",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "termcolor>=2.5.0",
    "torch>=2.6.0",
    "transformers>=4.50.3",
    "urllib3>=2.3.0",
    "uvicorn>=0.34.0",
]
